+++ src/edcon/utils/logging.py
@@ -6,7 +6,15 @@ from rich.logging import RichHandler
 class Logging:
     """Class that contains common functions for logging."""
 
+    logger = logging.getLogger('edcon')
     def __init__(self, logging_level=logging.INFO, filename=None):
+        logging.basicConfig(format='%(message)s',
+                            datefmt="[%X]",
+                            handlers=[RichHandler()])
+
+        Logging.logger.setLevel(logging_level)
+        Logging.logger.propagate = False
+
         if filename:
             self.enable_file_logging(filename, logging_level)
         else:
@@ -14,15 +22,17 @@ class Logging:
 
     def enable_stream_logging(self, logging_level):
         """Enables logging to stream using the provided log level with rich log formatting."""
-        logging.basicConfig(level=logging_level,
-                            format='%(message)s',
-                            datefmt="[%X]",
-                            handlers=[RichHandler()])
+        handler = RichHandler()
+        handler.setLevel(logging_level)
+        formatter = logging.Formatter(fmt='%(message)s', datefmt="[%X]")
+        handler.setFormatter(formatter)
+        Logging.logger.addHandler(handler)
 
     def enable_file_logging(self, filename, logging_level):
         """Enables logging to a file using the provided filename and log level."""
-        self.filename = filename
-        logging.basicConfig(filename=filename,
-                            level=logging_level,
-                            format='%(message)s',
-                            datefmt="[%X]")
+        handler = logging.FileHandler(filename)
+        handler.setLevel(logging_level)
+        formatter = logging.Formatter(fmt="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
+                                      datefmt="[%X]")
+        handler.setFormatter(formatter)
+        Logging.logger.addHandler(handler)