+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Patch] Run branch pipeline only if no MR pipeline exists.
 - [Patch] Upgraded pnu_map.csv.
 - [Patch] ComEthernetip: increased waiting time for send_io to ensure that data is sent reliably.
+- [Patch] Using dedicated logger instead of generic root logger for edcon logs.
 
 ## v0.10.7 - 21.07.23
 ### Added